create tests for barcodereader

+++ src/main/java/org/example/Car.java
@@ -21,7 +21,7 @@ public class Car {
         this.plate = plate;
 
     }
-    public Car(int barcode){
+    public Car(long barcode){
         this.barcode = String.valueOf(barcode);
     }
     public String getPlate(){return this.plate;}
